+++ docs/site/.vitepress/theme/index.js
@@ -1,0 +1,9 @@
+import DefaultTheme from 'vitepress/theme'
+import DownloadLinks from './DownloadLinks.vue'
+
+export default {
+  extends: DefaultTheme,
+  enhanceApp({ app }) {
+    app.component('DownloadLinks', DownloadLinks)
+  }
+}